Create playGame function
The function will take in the row and column of the square the player selects, it will display that selection to the player and place the active players symbol there if the move is valid. Each turn checkForWinner will check if there is a winner if there is the game will end, if not the active player will switch and the next round starts.