Create checkForWinner Function
The function creates a variable to hold the array of boardValues, it creates the array that holds the 8 possible sets of coordinates to win from that variable. It then checks if one set of coords has all 3 of their coordinates equal to the active players symbol which is passed to this function when it is called.