Create a working 'restart' button
The restartHandler first sets the values of all the squares in the array back to 0 and then resets the active player. If the 'click' event has been removed if a win or tie occured it is added back and then the boardDisplay and gameState are updated to display the correct information.